+++ react/components/FilterOptionTemplate.js
@@ -55,12 +55,12 @@ export default class FilterOptionTemplate extends Component {
       return null
     }
 
-    const className = classNames('vtex-search-result__filter pv3 bb b--light-gray', {
+    const className = classNames('vtex-search-result__filter pv3 bb b--muted-4', {
       'vtex-search-result__filter--selected': selected,
       'vtex-search-result__filter--available': !selected,
     })
 
-    const titleClassName = classNames('vtex-search-result__filter-title f6 flex items-center justify-between', {
+    const titleClassName = classNames('vtex-search-result__filter-title t-heading-7 flex items-center justify-between', {
       'ttu': selected,
     })