+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update loading to use an overlay instead of content loader after initial mount.
 
 ### Changed
-- Replace tachyons colors with design tokens.
+- Replace tachyons classes with design tokens.
 
 ### Added
 - Not found page component.