+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- Replace tachyons classes with design tokens.
 - Paddings and margins to match other components.
+- Update loading to use an overlay instead of content loader after initial mount.
+
+### Added
+- Not found page component.
 
 ## [2.4.1] - 2018-11-22
 ### Changed