+++ CHANGELOG.md
@@ -7,10 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- Update loading to use an overlay instead of content loader after initial mount.
-
-### Changed
 - Replace tachyons classes with design tokens.
+- Paddings and margins to match other components.
+- Update loading to use an overlay instead of content loader after initial mount.
 
 ### Added
 - Not found page component.