+++ CHANGELOG.md
@@ -7,12 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Add new mobile design.
 - Not found page component.
-- Paddings and margins to match other components.
 
 ### Changed
-- Update loading to use an overlay instead of content loader after initial mount.
 - Replace tachyons classes with design tokens.
+- Paddings and margins to match other components.
+- Update loading to use an overlay instead of content loader after initial mount.
 
 ## [2.4.1] - 2018-11-22
 ### Changed